src/main/java/com/outgrowthsolutions/ogsrecipeapp/converters/UnitOfMeasureCommandToUnitOfMeasure.java
package com.outgrowthsolutions.ogsrecipeapp.converters;

import com.outgrowthsolutions.ogsrecipeapp.commands.UnitOfMeasureCommand;
import com.outgrowthsolutions.ogsrecipeapp.domain.UnitOfMeasure;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;

@Component
public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand, UnitOfMeasure> {
    @Override
    public UnitOfMeasure convert(UnitOfMeasureCommand unitOfMeasureCommand) {
        if (unitOfMeasureCommand == null)
            return null;

        final UnitOfMeasure converted = new UnitOfMeasure();
        converted.setId(unitOfMeasureCommand.getId());
        converted.setDescription(unitOfMeasureCommand.getDescription());
        return converted;

    }
}
